# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wisetsuk Nakorn Condo Booking API
  version: 1.0.0
  description: |
    REST API ระบบจองห้องพักวิเศษสุขนครคอนโด
    แหล่งความจริงเดียวของ contract — ทุก endpoint ตรงกับหัวข้อ API Contract ใน docs/SPEC.md
servers:
  - url: http://localhost:8080
    description: development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: หมุน refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: ได้ token คู่ใหม่
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/me:
    get:
      tags: [me]
      summary: ข้อมูลผู้ใช้ที่ล็อกอินอยู่
      operationId: getMe
      responses:
        '200':
          description: ข้อมูลผู้ใช้
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [me]
      summary: แก้ไขโปรไฟล์
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: แก้ไขสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/me/password:
    post:
      tags: [me]
      summary: เปลี่ยนรหัสผ่าน (เพิกถอน refresh token ทุกใบ)
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: เปลี่ยนรหัสผ่านสำเร็จ
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /api/v1/me/notifications/read:
    post:
      tags: [me]
      summary: ทำเครื่องหมายว่าอ่านแล้ว (ไม่ระบุ id = อ่านทั้งหมด)
      operationId: markNotificationsRead
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  pattern: "^ntf-[0-9]+$"
      responses:
        '204':
          description: สำเร็จ ตอบ 204 แม้ id ไม่ใช่ของตัวเองหรือไม่มีจริง
...
  /api/v1/admin/members:
    get:
      tags: [admin]
      summary: รายชื่อสมาชิก
      operationId: listMembers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: รายชื่อสมาชิกแบบแบ่งหน้า
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
...
    Unauthorized:
      description: ไม่ได้ล็อกอินหรือ token ใช้ไม่ได้
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: กรุณาเข้าสู่ระบบ
...
    ValidationFailed:
      description: ข้อมูลไม่ผ่านการตรวจสอบ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_failed
              message: ข้อมูลไม่ถูกต้อง
              fields:
                email: อีเมลนี้ถูกใช้งานแล้ว
  schemas:
    ErrorCode:
      type: string
      enum:
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - bad_request
        - validation_failed
        - method_not_allowed
        - internal_error
        - invalid_credentials
        - account_disabled
        - room_unavailable
        - invalid_state
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: ข้อความภาษาไทยสำหรับผู้ใช้
            fields:
              type: object
              nullable: true
              description: มีเฉพาะตอน code = validation_failed
              additionalProperties:
                type: string
...
    UserRole:
      type: string
      description: >
        ตรงกับ enum user_role ใน PostgreSQL — สะกดติดกันไม่มีขีดล่าง
        frontend ใช้ค่านี้ตัดสินใจว่าจะพาผู้ใช้ไปหน้าไหนหลังล็อกอิน
        (member → หน้าสมาชิก, admin → หน้าผู้ดูแล, superadmin → หน้าหัวหน้าผู้ดูแล)
      enum: [member, admin, superadmin]
...
    PageMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    User:
      type: object
      required: [id, email, first_name, last_name, phone, role, is_active, created_at]
      properties:
        id:
          type: string
          pattern: "^usr-[0-9]+$"
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        branch_id:
          type: string
          pattern: "^brn-[0-9]+$"
          nullable: true
          description: >
            สาขาที่รับผิดชอบ มีค่าเฉพาะ role = admin ซึ่งผูกกับสาขาเดียวเสมอ
            (superadmin เข้าถึงทุกสาขาอยู่แล้วจึงไม่ผูกกับสาขาใด)
        branch_name:
          type: string
          description: ชื่อของสาขาข้างบน ส่งมาด้วยเพื่อให้ตารางแสดงผลได้โดยไม่ต้องยิงถามรายการสาขาซ้ำ
        must_change_password:
          type: boolean
          description: >
            true เมื่อบัญชีนี้ยังใช้รหัสผ่านตั้งต้นที่ระบบกำหนดให้ตอนสร้าง
            frontend ต้องบังคับให้ตั้งรหัสผ่านใหม่ก่อนใช้งานส่วนอื่น
        avatar_url:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    TokenPair:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        phone:
          type: string
          pattern: '^[0-9]{9,10}$'
        avatar_url:
          type: string
          format: uri
          nullable: true
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password, confirm_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 72
        confirm_password:
          type: string
...